Move exception classes to the common package
COSDK-500
