+++ core/api/core.api
@@ -148,6 +148,18 @@ public final class com/adyen/checkout/core/analytics/internal/data/remote/model/
 	public synthetic fun newArray (I)[Ljava/lang/Object;
 }
 
+public final class com/adyen/checkout/core/common/exception/HttpException : java/lang/RuntimeException {
+	public static final field $stable I
+	public fun <init> (ILjava/lang/String;Lcom/adyen/checkout/core/common/internal/model/ErrorResponseBody;)V
+	public final fun getCode ()I
+	public final fun getErrorBody ()Lcom/adyen/checkout/core/common/internal/model/ErrorResponseBody;
+}
+
+public final class com/adyen/checkout/core/common/exception/ModelSerializationException : java/lang/RuntimeException {
+	public static final field $stable I
+	public fun <init> (Ljava/lang/Class;Lorg/json/JSONException;)V
+}
+
 public final class com/adyen/checkout/core/common/internal/model/EmptyResponse$Companion {
 }
 
@@ -781,18 +793,6 @@ public final class com/adyen/checkout/core/components/paymentmethod/PaymentMetho
 	public final fun getSUPPORTED_ACTION_ONLY_PAYMENT_METHODS ()Ljava/util/List;
 	public final fun getSUPPORTED_PAYMENT_METHODS ()Ljava/util/List;
 	public final fun getUNSUPPORTED_PAYMENT_METHODS ()Ljava/util/List;
-}
-
-public final class com/adyen/checkout/core/exception/HttpException : java/lang/RuntimeException {
-	public static final field $stable I
-	public fun <init> (ILjava/lang/String;Lcom/adyen/checkout/core/common/internal/model/ErrorResponseBody;)V
-	public final fun getCode ()I
-	public final fun getErrorBody ()Lcom/adyen/checkout/core/common/internal/model/ErrorResponseBody;
-}
-
-public final class com/adyen/checkout/core/exception/ModelSerializationException : java/lang/RuntimeException {
-	public static final field $stable I
-	public fun <init> (Ljava/lang/Class;Lorg/json/JSONException;)V
 }
 
 public abstract class com/adyen/checkout/core/internal/ActionHandlingPaymentMethodConfigurationBuilder : com/adyen/checkout/core/internal/BaseConfigurationBuilder, com/adyen/checkout/core/internal/ActionHandlingConfigurationBuilder {
