+++ core/src/main/java/com/adyen/checkout/core/analytics/internal/data/remote/model/AnalyticsSetupRequest.kt
@@ -8,6 +8,7 @@
 
 package com.adyen.checkout.core.analytics.internal.data.remote.model
 
+import com.adyen.checkout.core.common.exception.ModelSerializationException
 import com.adyen.checkout.core.common.internal.model.JsonUtils
 import com.adyen.checkout.core.common.internal.model.ModelObject
 import com.adyen.checkout.core.common.internal.model.ModelUtils
@@ -16,7 +17,6 @@ import com.adyen.checkout.core.common.internal.model.getLongOrNull
 import com.adyen.checkout.core.common.internal.model.getStringOrNull
 import com.adyen.checkout.core.common.internal.model.optStringList
 import com.adyen.checkout.core.components.data.model.Amount
-import com.adyen.checkout.core.exception.ModelSerializationException
 import kotlinx.parcelize.Parcelize
 import org.json.JSONException
 import org.json.JSONObject
