+++ core/src/main/java/com/adyen/checkout/core/analytics/internal/data/remote/model/AnalyticsTrackError.kt
@@ -8,10 +8,10 @@
 
 package com.adyen.checkout.core.analytics.internal.data.remote.model
 
+import com.adyen.checkout.core.common.exception.ModelSerializationException
 import com.adyen.checkout.core.common.internal.model.ModelObject
 import com.adyen.checkout.core.common.internal.model.getLongOrNull
 import com.adyen.checkout.core.common.internal.model.getStringOrNull
-import com.adyen.checkout.core.exception.ModelSerializationException
 import kotlinx.parcelize.Parcelize
 import org.json.JSONException
 import org.json.JSONObject
