+++ core/src/main/java/com/adyen/checkout/core/analytics/internal/data/remote/model/AnalyticsTrackRequest.kt
@@ -8,11 +8,11 @@
 
 package com.adyen.checkout.core.analytics.internal.data.remote.model
 
+import com.adyen.checkout.core.common.exception.ModelSerializationException
 import com.adyen.checkout.core.common.internal.model.ModelObject
 import com.adyen.checkout.core.common.internal.model.ModelUtils
 import com.adyen.checkout.core.common.internal.model.ModelUtils.deserializeOptList
 import com.adyen.checkout.core.common.internal.model.getStringOrNull
-import com.adyen.checkout.core.exception.ModelSerializationException
 import kotlinx.parcelize.Parcelize
 import org.json.JSONException
 import org.json.JSONObject
