+++ core/src/main/java/com/adyen/checkout/core/common/internal/api/OkHttpClient.kt
@@ -8,9 +8,9 @@
 
 package com.adyen.checkout.core.common.internal.api
 
+import com.adyen.checkout.core.common.exception.HttpException
+import com.adyen.checkout.core.common.exception.ModelSerializationException
 import com.adyen.checkout.core.common.internal.model.ErrorResponseBody
-import com.adyen.checkout.core.exception.HttpException
-import com.adyen.checkout.core.exception.ModelSerializationException
 import okhttp3.Headers.Companion.toHeaders
 import okhttp3.HttpUrl.Companion.toHttpUrlOrNull
 import okhttp3.MediaType.Companion.toMediaType
