+++ core/src/main/java/com/adyen/checkout/core/common/internal/model/ErrorResponseBody.kt
@@ -9,7 +9,7 @@
 package com.adyen.checkout.core.common.internal.model
 
 import androidx.annotation.RestrictTo
-import com.adyen.checkout.core.exception.ModelSerializationException
+import com.adyen.checkout.core.common.exception.ModelSerializationException
 import kotlinx.parcelize.Parcelize
 import org.json.JSONException
 import org.json.JSONObject
