+++ core/src/main/java/com/adyen/checkout/core/components/paymentmethod/MBWayPaymentMethod.kt
@@ -8,8 +8,8 @@
 
 package com.adyen.checkout.core.components.paymentmethod
 
+import com.adyen.checkout.core.common.exception.ModelSerializationException
 import com.adyen.checkout.core.common.internal.model.getStringOrNull
-import com.adyen.checkout.core.exception.ModelSerializationException
 import kotlinx.parcelize.Parcelize
 import org.json.JSONException
 import org.json.JSONObject
